Fix bug dropping 'changes' from summary on some events
When events have a name where the 2nd and 3rd parameters of the name are equal, changes where not being checked and processed for deeper info. This resulted in some potentially useful information being dropped from the summary output. Corrected the problem so that this does not occur.

Additionally improved the script a small amount so that commas are output into initial json only between events, instead of after every one. This allows the final trailing 0 to be skipped and the output to be cleaner.

Corrected an additional minor bug in that a "line" of initial XML input could be detected as the end of a DB row even when no opening tag ever occured. This would never occur in properly formatted XML but the condition should be checked none the less.

Also copied the output of the line number reference into the section handling 2nd and 3rd paramters of event name being equal. It was missing.